fix(js): datetime now can compile when JS

+++ src/pylib/Lib/datetime_impl/delta_chk.nim
@@ -8,7 +8,7 @@ func outOfDay(delta: timedelta): bool =
   # NIM-BUG: `convert(Days, Microseconds, 1)`:
   # times.nim(417, 65)
   # Error: illegal conversion from '86400000000' to '[-2147483648..2147483647]'
-  const OneDayMs = 1_000_000 * 3600 * 24
+  const OneDayMs = int64(1_000_000) * 3600 * 24
   abs(delta.inMicroseconds) > OneDayMs
 
 template chkOneDay*(delta: timedelta) =
